v812 Fixes, master branch (2024.09.05.)

See merge request atomki-nuclear-phys/cda!5

+++ devices/v812/ChannelConfig.cxx
@@ -74,7 +74,8 @@ namespace v812 {
       bool ok;
       m_channel = element.attribute( "ChannelNumber", "" ).toInt( &ok );
       CHECK( ok );
-      m_threshold = element.attribute( "Threshold", "1" ).toInt( &ok );
+      m_threshold =
+         static_cast<uint8_t>(element.attribute( "Threshold", "1" ).toUShort( &ok ));
       CHECK( ok );
 
       // Print the configuration for debugging.
@@ -91,7 +92,8 @@ namespace v812 {
 
       // Write the channel's configuration.
       element.setAttribute( "ChannelNumber", m_channel );
-      element.setAttribute( "Threshold", m_threshold );
+      element.setAttribute( "Threshold",
+                            static_cast<unsigned int>(m_threshold) );
 
       // Return gracefully.
       return StatusCode::SUCCESS;
@@ -108,12 +110,12 @@ namespace v812 {
       return;
    }
 
-   uint16_t ChannelConfig::getThreshold() const {
+   uint8_t ChannelConfig::getThreshold() const {
 
       return m_threshold;
    }
 
-   void ChannelConfig::setThreshold( uint16_t value ) {
+   void ChannelConfig::setThreshold( uint8_t value ) {
 
       m_threshold = value;
       return;
